Safari-609.1.11.1/LayoutTests/http/tests/notifications/window-show-on-click-expected.txt

This test checks that a script can open windows when the user clicks on a native notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Notification.permission is "granted"
PASS clicking on notification successfully opened a window
PASS successfullyParsed is true

TEST COMPLETE
